requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai-whisper==20240930
python-multipart==0.0.9